Isolate changes in separate file.
